refactor blog revision list as bfdslist
also updated BfDsListItem to support more advanced interactions


Summary:

Test Plan:
