+++ packages/bfDs/components/BfDsListItem.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { forwardRef, useImperativeHandle, useState } from "react";
 import {
   BfDsIcon,
   type BfDsIconType,
@@ -6,10 +6,16 @@ import {
 import { classnames } from "lib/classnames.ts";
 import { BfDsToggle } from "packages/bfDs/components/BfDsToggle.tsx";
 
+export type BfDsListItemHandle = {
+  setExpand: (value: boolean) => void;
+  isExpanded: () => boolean;
+};
+
 type Props = {
   action?: React.ReactNode;
   content: string | React.ReactNode;
   expandedContent?: React.ReactNode;
+  expandCallback?: (newExpandingState?: boolean) => void;
   iconLeft?: BfDsIconType;
   iconLeftColor?: string;
   iconRight?: BfDsIconType;
@@ -22,98 +28,113 @@ type Props = {
   xstyle?: React.CSSProperties;
 };
 
-export function BfDsListItem(
-  {
-    action,
-    content,
-    expandedContent,
-    iconLeft,
-    iconLeftColor,
-    iconRight,
-    isHighlighted,
-    footer,
-    onClick,
-    onDoubleClick,
-    toggle,
-    toggled,
-    xstyle,
-  }: Props,
-) {
-  const [expand, setExpand] = useState(false);
-  function handleClick(e: React.MouseEvent<HTMLDivElement, MouseEvent>) {
-    if ((e.target as HTMLElement).closest(".ignore-internal-click")) {
-      return;
-    }
-    if (expandedContent) {
-      setExpand(!expand);
-    }
-    if (onClick) {
-      onClick();
-    }
-  }
-  const clickable = (typeof onClick === "function" && !isHighlighted) ||
-    expandedContent != null;
+export const BfDsListItem = forwardRef<BfDsListItemHandle, Props>(
+  (
+    {
+      action,
+      content,
+      expandedContent,
+      expandCallback,
+      iconLeft,
+      iconLeftColor,
+      iconRight,
+      isHighlighted,
+      footer,
+      onClick,
+      onDoubleClick,
+      toggle,
+      toggled,
+      xstyle,
+    },
+    ref,
+  ) => {
+    const [expand, setExpand] = useState(false);
 
-  const listItemClasses = classnames([
-    "list-item",
-    { isHighlighted },
-    { clickable },
-  ]);
+    useImperativeHandle(ref, () => ({
+      setExpand: (value: boolean) => {
+        setExpand(value);
+        expandCallback?.(value);
+      },
+      isExpanded: () => expand,
+    }));
 
-  return (
-    <div
-      className={listItemClasses}
-      onClick={(e) => handleClick(e)}
-      style={xstyle}
-      onDoubleClick={onDoubleClick}
-    >
-      <div className="list-item-row">
-        {iconLeft && (
-          <div className="list-item-icon">
-            <BfDsIcon
-              name={iconLeft}
-              color={iconLeftColor ?? "var(--textSecondary)"}
-            />
+    function handleClick(e: React.MouseEvent<HTMLDivElement, MouseEvent>) {
+      if ((e.target as HTMLElement).closest(".ignore-internal-click")) {
+        return;
+      }
+      const currendExpandedState = expand;
+      if (expandedContent) {
+        setExpand(!currendExpandedState);
+        expandCallback?.(!currendExpandedState);
+      }
+      if (onClick) {
+        onClick();
+      }
+    }
+    const clickable = (typeof onClick === "function" && !isHighlighted) ||
+      expandedContent != null;
+
+    const listItemRowClasses = classnames([
+      "list-item-row",
+      { isHighlighted },
+      { clickable },
+    ]);
+
+    return (
+      <div
+        className="list-item"
+        style={xstyle}
+        onClick={(e) => handleClick(e)}
+        onDoubleClick={onDoubleClick}
+      >
+        <div className={listItemRowClasses}>
+          {iconLeft && (
+            <div className="list-item-icon">
+              <BfDsIcon
+                name={iconLeft}
+                color={iconLeftColor ?? "var(--textSecondary)"}
+              />
+            </div>
+          )}
+
+          <div className="list-item-main">
+            <div className="list-item-text">{content}</div>
+            {footer && (
+              <div className="list-item-meta">
+                {footer}
+              </div>
+            )}
           </div>
-        )}
-
-        <div className="list-item-main">
-          <div className="list-item-text">{content}</div>
-          {footer && (
-            <div className="list-item-meta">
-              {footer}
+          {expandedContent && (
+            <div className="list-item-icon">
+              <BfDsIcon
+                name={expand ? "arrowUp" : "arrowDown"}
+                color="var(--textSecondary)"
+              />
+            </div>
+          )}
+          {iconRight && !expandedContent && (
+            <div className="list-item-icon">
+              <BfDsIcon name={iconRight} color="var(--textSecondary)" />
+            </div>
+          )}
+          {toggle && (
+            <div className="list-item-toggle">
+              <BfDsToggle value={!!toggled} onChange={toggle} />
+            </div>
+          )}
+          {action && (
+            <div className="list-item-action ignore-internal-click">
+              {action}
             </div>
           )}
         </div>
-        {expandedContent && (
-          <div className="list-item-icon">
-            <BfDsIcon
-              name={expand ? "arrowUp" : "arrowDown"}
-              color="var(--textSecondary)"
-            />
-          </div>
-        )}
-        {iconRight && !expandedContent && (
-          <div className="list-item-icon">
-            <BfDsIcon name={iconRight} color="var(--textSecondary)" />
-          </div>
-        )}
-        {toggle && (
-          <div className="list-item-toggle">
-            <BfDsToggle value={!!toggled} onChange={toggle} />
-          </div>
-        )}
-        {action && (
-          <div className="list-item-action ignore-internal-click">
-            {action}
+        {expandedContent && expand && (
+          <div className="list-item-expanded ignore-internal-click">
+            {expandedContent}
           </div>
         )}
       </div>
-      {expandedContent && expand && (
-        <div className="list-item-expanded">
-          {expandedContent}
-        </div>
-      )}
-    </div>
-  );
-}
+    );
+  },
+);
